Update pagination logic, search query and update pagination by result search
